requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
google-auth==2.49.1
requests==2.33.1
